Implement custom software list loading
Software specified on the command line with an image format now looks for a file named 'software.xml' in the software's subdirectory or ZIP file if the driver-configured softlists have no info on it. This XML file should have the usual MAME software list format, and must contain an entry with the same name as the software; other entries are normally ignored.

Parsed software lists are now separate from devices, allowing them to exist independently from the configured state.

Tighten restrictions on "requirement" feature in software entries to simplify loading
- Required additional software must specify the software list it should be loaded from unless it can be found in the same list (which usually never happens, though it can happen with custom software lists). The validity checker now enforces this.
- Required additional software cannot be automatically loaded by any part other than the first.
